# Application_Samples/Todo_App/TodoApp_Client/Blz_Web/Models/Todo.cs
namespace Blz_Web.Models;

public class Todo
{
    public int? Id { get; set; }
    public string? Task { get; set; }
    public int? AssignTo { get; set; }
    public bool? IsComplete { get; set; }
}